Started work on using GPU again
Former-commit-id: 60b000bc091b1656a13169a75865decc701d6227

+++ argumentparser.hpp
@@ -16,5 +16,5 @@ struct ArgumentParser
 	std::string outfile { "" };
 	std::vector<int> NamedColumns {{1,2}};
 	bool should_exit { false };
-	size_t memoryAvailable{0};
+	long memoryAvailable{0};
 };
